summarizer_gc_gemma4_31b/pyproject.toml
[project]
name = "summarizer-gc-gemma4-31b"
version = "0.1.0"
description = "Google Cloud Gemma 4 31B specific summarizer"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "pyyaml",
]

[project.scripts]
summarizer-gc-gemma4-31b = "summarizer_gc_gemma4_31b:main"
